Plagiarism prevention changes
Tweets on this coursework have a very specific name; to prevent indexing by search engines, I have adjusted this on this commit. Please see the README.

Note that previously, a basic implementation was provided for the Tweet class (under its older, identifiable name). This was purely to allow the comparator implementation to compile without providing access to the proprietary class. This has now been removed as it delivered undefined behaviour and was there merely to ensure compilation. In the actual use, that class operated very differently.